not always delete root files

+++ clean.sh
@@ -1,3 +1,6 @@
-rm *.root
 rm *~
 rm *cfg.py
+
+if [ $1 = "r" ]; then
+    rm *.root
+fi
